refactor(dashboard): rename AgentConfig to HermeumConfig
