+++ apps/dashboard/src/entities/index.ts
@@ -1,5 +1,5 @@
 export * from "./auth";
 export * from "./agent";
 export * from "./template";
-export * from "./agent-config";
+export * from "./hermeum-config";
 export * from "./secret";
