+++ apps/dashboard/src/server/infras/local-hermeum-config.ts
@@ -1,12 +1,12 @@
 import * as fs from "node:fs";
 import { parse } from "yaml";
 
-import { AgentConfig, AgentConfigSchema } from "@/entities";
+import { HermeumConfig, HermeumConfigSchema } from "@/entities";
 import { config } from "@/server/libs/config";
 import { ConfigAdaptor } from "../usecases/adaptors/config";
 
 export class LocalConfig implements ConfigAdaptor {
-  private cache: AgentConfig;
+  private cache: HermeumConfig;
 
   constructor(private filePath?: string) {
     const resolvedPath = this.filePath ?? config.agentConfigPath;
@@ -14,17 +14,17 @@ export class LocalConfig implements ConfigAdaptor {
     try {
       const content = fs.readFileSync(resolvedPath, "utf-8");
       const raw = parse(content);
-      this.cache = AgentConfigSchema.parse(raw);
+      this.cache = HermeumConfigSchema.parse(raw);
     } catch (err) {
       if ((err as NodeJS.ErrnoException).code === "ENOENT") {
-        this.cache = AgentConfigSchema.parse({ templates: [] });
+        this.cache = HermeumConfigSchema.parse({ templates: [] });
       } else {
         throw err;
       }
     }
   }
 
-  get(): AgentConfig {
+  get(): HermeumConfig {
     return this.cache;
   }
 }
