+++ apps/dashboard/src/server/usecases/agent.ts
@@ -13,7 +13,7 @@ import {
 } from "@/entities";
 
 import { KubernetesClient } from "../infras/kubernetes/client";
-import { LocalConfig } from "../infras/local-agent-config";
+import { LocalConfig } from "../infras/local-hermeum-config";
 import { Runtime } from "./adaptors/runtime";
 import { ConfigAdaptor } from "./adaptors/config";
